Project Colors in Project Editor
